+++ education_theme/__manifest__.py
@@ -1,48 +1,47 @@
 # -*- coding: utf-8 -*-
 ###############################################################################
-#    A part of Educational ERP Project <https://www.educationalerp.com>
 #
 #    Cybrosys Technologies Pvt. Ltd.
-#    Copyright (C) 2021-TODAY Cybrosys Technologies (<https://www.cybrosys.com>)
-#    Author: Hajaj Roshan (hajaj@cybrosys.in)
 #
-#    This program is free software: you can modify
-#    it under the terms of the GNU Affero General Public License (AGPL) as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
+#    Copyright (C) 2023-TODAY Cybrosys Technologies(<https://www.cybrosys.com>)
+#    Author: Raneesha M K (odoo@cybrosys.com)
+#
+#    You can modify it under the terms of the GNU AFFERO
+#    GENERAL PUBLIC LICENSE (AGPL v3), Version 3.
 #
 #    This program is distributed in the hope that it will be useful,
 #    but WITHOUT ANY WARRANTY; without even the implied warranty of
 #    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
+#    GNU AFFERO GENERAL PUBLIC LICENSE (AGPL v3) for more details.
 #
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#    You should have received a copy of the GNU AFFERO GENERAL PUBLIC LICENSE
+#    (AGPL v3) along with this program.
+#    If not, see <http://www.gnu.org/licenses/>.
 #
 ###############################################################################
 {
     'name': 'Educational ERP Theme',
-    'version': '16.0.1.0.0',
-    'summary': """Insight to Educational ERP""",
-    'description': """In trio combination of red, green and yellow Educational ERP,
-    opens the new gate way for management""",
-    'category': 'Theme',
+    'version': '16.0.2.0.0',
+    'summary': """Attractive and unique red-teal combination of backend-end 
+                    theme for Education ERP""",
+    'description': """In trio combination of red, teal and white Educational 
+                 ERP opens the new gate way for management""",
+    "category": "Theme/Backend",
     'author': 'Cybrosys Techno Solutions',
     'company': 'Cybrosys Techno Solutions',
     'maintainer': 'Cybrosys Techno Solutions',
     'website': "https://www.educationalerp.com",
-    'depends': ['web'],
+    'depends': ['base'],
     'data': [
-        'views/education_theme.xml',
+        'data/logo_data.xml',
+        'views/shortcut_icon_templates.xml',
     ],
     'assets': {
         'web.assets_backend': [
-            '/education_theme/static/src/css/education_backend.css'
+            'education_theme/static/src/scss/education_backend.scss',
+            'education_theme/static/src/js/sidebar_menu.js',
+            'education_theme/static/src/xml/sidebar.xml'
         ],
-        'web.assets_frontend': [
-            '/education_theme/static/src/css/education_frontend.css'
-        ],
-
     },
     'images': [
         'static/description/banner.png',